+++ CHANGELOG.md
@@ -84,6 +84,11 @@ The change log for the Rust [socketcan](https://crates.io/crates/socketcan) libr
 - `CanInterface::set_data_bitrate()` now has the same debug-build sanity checks as `set_bitrate()` (bitrate and sample-point range), with an FD-appropriate upper bound of 8 Mbit/s for the data phase
 
 
+## [Version 3.6.2](https://github.com/socketcan-rs/socketcan-rs/compare/v3.6.1..v3.6.2)  (2026-06-19)
+
+- [#103](https://github.com/socketcan-rs/socketcan-rs/pull/103) Disable async-io/async-std/smol features for docs.rs
+
+
 ## [Version 3.6.1](https://github.com/socketcan-rs/socketcan-rs/compare/v3.6.0..v3.6.1)  (2026-06-10)
 
 - [#101](https://github.com/socketcan-rs/socketcan-rs/pull/101) Add libc::ioctl fix for musl targets